+++ app/Http/Controllers/Admin/EffectivenesController.php
@@ -54,6 +54,10 @@ class EffectivenesController extends Controller
     public function destroy($id): RedirectResponse
     {
         $item = Effectivenes::findOrFail($id);
+        if ($item->bookings->count()) {
+            flash(__('effectivenes.messages.has_bookings'))->error();
+            return to_route($this->route . '.index');
+        }
         if ($item->delete()) {
             flash(__('effectivenes.messages.deleted'))->success();
         }
