+++ app/Http/Controllers/Admin/TripController.php
@@ -56,6 +56,10 @@ class TripController extends Controller
     public function destroy($id): RedirectResponse
     {
         $item = Trip::findOrFail($id);
+        if ($item->bookings->count()) {
+            flash(__('trips.messages.has_bookings'))->error();
+            return to_route($this->route . '.index');
+        }
         if ($item->delete()) {
             flash(__('trips.messages.deleted'))->success();
         }
